+++ relay-dynamic-config/src/global.rs
@@ -168,8 +168,6 @@ pub struct Options {
     pub feedback_ingest_topic_rollout_rate: f32,
 
     /// Flag for handling feedback and attachments in the same envelope.
-    ///
-    /// Enabling this flag will also produce feedback to Kafka as UserReportV2KafkaMessage
     #[serde(
         rename = "feedback.ingest-inline-attachments",
         deserialize_with = "default_on_error",